pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "safco-dental-scraper"
version = "0.1.0"
description = "AI agent-based scraping system for Safco Dental Supply"
requires-python = ">=3.11"
dependencies = [
    "httpx",
    "beautifulsoup4",
    "lxml",
    "pydantic",
    "pyyaml",
    "anthropic",
    "rich",
    "aiosqlite",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
]
